services/admin-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
asyncpg==0.29.0
aiosqlite==0.19.0
httpx==0.25.0
aiohttp==3.9.0
python-jose[cryptography]==3.3.0
requests==2.31.0
aiofiles==23.2.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pandas==2.1.3
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
reportlab==4.0.6
jinja2==3.1.2
email-validator==2.1.0
celery==5.3.4
redis==5.0.1
